+++ lib/theme/icons.dart
@@ -16,7 +16,10 @@ class AIcons {
   static const IconData checked = Icons.done_outlined;
   static const IconData counter = Icons.plus_one_outlined;
   static const IconData date = Icons.calendar_today_outlined;
+  static const IconData dateRecent = Icons.today_outlined;
+  static const IconData dateUndated = Icons.event_busy_outlined;
   static const IconData description = Icons.description_outlined;
+  static const IconData descriptionUntitled = Icons.comments_disabled_outlined;
   static const IconData disc = Icons.fiber_manual_record;
   static const IconData display = Icons.light_mode_outlined;
   static const IconData error = Icons.error_outline;
@@ -35,7 +38,6 @@ class AIcons {
   static const IconData ratingRejected = MdiIcons.starMinusOutline;
   static const IconData ratingUnrated = MdiIcons.starOffOutline;
   static const IconData raw = Icons.raw_on_outlined;
-  static const IconData recent = Icons.today_outlined;
   static const IconData shooting = Icons.camera_outlined;
   static const IconData removableStorage = Icons.sd_storage_outlined;
   static const IconData sensorControlEnabled = Icons.explore_outlined;